+++ apps/desktop-solid/src/components/Header.tsx
@@ -1,7 +1,7 @@
 export default function Header() {
   return (
     <div
-      class="flex items-center justify-start space-x-1 h-[44px] shrink-0 px-3 border-gray-200 border-b"
+      class="flex items-center justify-start bg-gray-50 space-x-1 h-[44px] shrink-0 px-3 border-gray-200 border-b"
       data-tauri-drag-region
     />
   );
