+++ packages/ui-solid/src/index.ts
@@ -1,4 +1,2 @@
-import "unfonts.css";
-
 export * from "./Button";
 export * from "./SwitchTab";
